Use already downloaded raw data for regression testing
This moves the data files for all but one dataset into the test
directory to save time on downloading and avoid issue with Travis
sometimes failing to download files.

nosetests needs to be run from inside the test directory.

Required addition of a setup fixture for the module to download
the scripts.