Force connection wizard to be large enough for 'choose file' button
Because (probably) app.ConnectPage.Draw contains different fields
depending on the engine, wx is not properly accomodating the
varying sizes of this page that might result. In cases where there
is a file attribute (sqlite, MS Access) the wizard is too small
on some systems and the button is cut off.

This is a hack that temporarily addresses this problem by adding
a large empty text box to the bottom of the first page of the
wizard. This probably should be fixed properly when time allows.