Use None for null values in json engine
This causes the json module to properly dump the json keyword null.
This is better than empty strings because some importers will treat
emptry strings as string values thus changing the data type for the
columns. It is better than removing the key:value pair because it
is explicit about the presence of the null.