Catch errors from jsonengine.disconnect to work with GUI
The disconnect method in the jsonengine is overloaded to allow
the proper formatting for the end of the JSON file to be written.
However, lib.app.connect_wizard.ConfirmPage forces disconnect to
be called before output_file is created. It does this for the
purpose of checking the connection, but since the output file
doesn't exist yet this causes an error.

We know catch that error since we simply want disconnect to pass
when the GUI is checking the connection for file based engines.