Use known locations for download only files when copying
An edge case bug existed where the file to be downloaded existed
in the search path, but not in the appropriate raw_data directory.
In this case the Retriever tried to copy the file from the
raw_data directory and threw an error. This situation could
occur if downloaded files are stored in version control, but
running the program calls the Retriever.

This is addressed by copying the file from the location where
find_file actually finds it. The details are reported so that
there is no confusion about where the files are coming from.