Allow explicitly listing files to be moved by the download engine
It is often desireable to download an archive with a number of
files to be extracted. The download_only engine would currently
only move those files if they were present in the raw_data
directory prior to running `retriever download DATASET` because
find_file didn't find them and therefore couldn't add them to
all_files.

This commit adds a register_files method that can be called from
a retriever script that processes an archive of files for download
to have the names of these files added to all_files following
the download.