Stem columns may be indicated by "dbh" as well as by "stem" in Gentry
Column headers vary from file to file in Gentry. We thought that
all of the stem column headers included the word "stem", but in
BEZA1 and BEZA3 (at least at the moment) they are indicated by
"SDBH". Since "dbh" doesn't occur elsewhere in the column headers
we now check this as well.