Make sure that CITATION is added to distributions
This adds a MANIFEST.in file to make sure that CITATION is included
in distributions. This should be handled by the data_files line in
setup.py, but that isn't working. This was apparently a bug in
Python 2.6 that was fixed, but there are reports of continued
issues in 2.7 http://stackoverflow.com/a/2998311/133513.