Manually fix incorrect column name in QUIAPACA gentry site
QUIAPACA has a column of counts, but it is mislabelled with the same header as the
DBH columns, STEMDBH. We have attempted to get the data provider to fix this and they
have not done so. This fix manually identifies the column correctly.

This change is fragile to an eventual fix by the data provider, but evidence over
the last several years suggests that such a fix is unlikely to occur.