Allow either wxPython 2.8 or 3.0 for Debian package
The current testing version of Debian only includes wxPython 3.0.
Since the name of the Debian package includes the version number this
means that the Debian package needs to allow 3.0 as well as 2.8 to
support both current and future Debian and Debian derived Linux distributions.

Thanks to @cboettig for pointing this out.

Fixes #247.