Add control of the scope of a reset
'retriever reset' now takes an argument controlling the scope of the reset.
This argument can be 'all' to remove all of the cached information or
'scripts', 'data', or 'connections' to only remove the named pieces.

Moves the reset logic into a function in lib/tools to keep it from cluttering up __main__.

Closes #215.