Check input path on download
A dataset can be downloaded to a specific path using -p argument.
But if the path provided doesn't exists, retriever creates a file with name same as path provided and overwrites it with each file of dataset.
This patch fixes it by creating the directory if it doesn't already exists, and displays error message if the directory cannot be created.
Fixes #205