On Windows have CLI default to storing data in directory retriever is run from
To avoid issues with the Installer installed version of the Retriever confusing folks
by defaulting to storing their data in the system folder where the Retriever is installed,
we had defaulted to storing all data on Windows to the Desktop. This is undesirable for the
CLI and for things that wrap the CLI like the ecoretriever R package. This keeps the Desktop
as the default when using the GUI on Windows, but returns the CLI behavior to the standard
of defaulting the to current working directory.