Only load a script the first time it is discovered
When MODULE_LIST() searches for scripts it looks from them in three places:
1. The working directory
2. The the 'scripts' directory in the working directory
3. '~/.retriever/scripts

It is possible for the same script to be present in multiple of these directories.
This is particularly true for developers who have the retriever installed and so
when developing open have copies of all scripts in both the repos scripts directory
and '~/.retriever/scripts'.

This change checks to see if a particular script has already been added to avoid
duplicates. The will prevent developers from getting duplicates in 'retriever ls',
from accidentally producing 'version.txt' with duplicate values (resulting in
duplicate downloads for users), and avoid tricky developer/testing bugs where the
wrong script is being used.