Allow adding a prefix to filenames determined from urls
In somecases the filenames generated from urls will not be particularly
informative and may not be unique or valid filenames. This is most
likely to occur when the file is coming from a webservice and the name
is not explicitly included in the url. This change allows an optional
prefix to be included for both regular files and archives.