Add option to keep extracted files in subdirectory
When extracting files from an archive the default behavior is to put them
all in a single directory for the dataset in the retriever's raw data directory.
This adds the ability to keep the files from each archive in a subdirectory
named after the archive. This makes it easier to place them in appropriate
subdirectories in final location when desired.