Add flag to keep downloaded data in subdirectory
Default behavior when using the download_only engine is that all files
are stored in the specified path, with the working directory as the default.
This change adds an optional `--subdir` flag to this engine that changes
this behavior to put the files in the subdirectory they were extracted to
in the raw data directory. This flag will be most useful in conjuction with
the new optional `keep_in_dir` argument in `download_files_from_archive`, to
allow multiple extracted archives for a single dataset to each be kept in there
own subdirectories.