Allow downloading archives without names in urls
When using web services to retrieve files the file names are sometimes not included in the url.
The retriever uses urls to create file names for raw data storage and access so this change
adds the ability to add a prefix to the "file name" that is being parsed from the url. In some webservices
this "file name" may just be the last parameter set and so require the prefix to be informative.

To solve this for archives this adds an optional archivename argument to download_files_from_archive
that allows a database script to provide the name for the file to be saved.