Update author and contact information in setup.py
Since Ben Morris is no longer actively developing this changes the contact
email to Ethan White's and add's Ethan as an author (which he has been all
along but we've been ignoring since PyPI doesn't really have a multi-authorship
concept.

This also chances the url to the github page since this is really about where
the package is not about the site for the home page.