Respect enclosing quotes when handling cross-tab data
The code for handling cross-tab data was not paying attention to
enclosing quotes, either when parsing the fields from the raw data or
when reassembling them. This meant that if there were fields that
included the delimiter, cross-tab manipulations resulted in incorrect
parsing of the data. This change reads and writes all cross-tab data
using the csv module to avoid these issues.

Solved by pair programming with @henrykironde.

Fixes #361.