Check for version 1.6 before importing DownloadOnlyTemplate in MammalSuperTree
If v1.6 is running use a dummy script to avoid importing a template that
doesn't exist in that version.

This will allow us to keep updated versions of all scripts available without breaking
existing installations of v1.6. This resulted because MASTER was set to TRUE accidentally
in the v1.6 release. See https://github.com/weecology/retriever/issues/208 for details.