Use non-raw_data directory for download regression tests
We now are checking for existing files before downloading to non-working directory
paths. Since the raw data for our regressions tests is stored in the same location
as the download tests were moving the data to, this prevented them from doing much.
This creates a new destination directory for the download based testing.