Add high speed bulk importing for SQLite
For files that do not need to be cleaned this adds high speed bulk importing
equivalent to that already available in MySQL and PostgreSQL.

This is accomplished by splitting the file up into intermediate sized chunks
(large enough to speed things up, small enough to consistently fit in memory)
and then inserting them using `executemany`.

This results in over a 10x speed up for BBS, and it's probably >20x since BBS
has several tables that can't be bulk inserted.