Update BBS scripts now that longitudes have proper signs
Longitudes in the BBS Routes tables used to be missing the negative sign.
This has been fixed in the most recent data release allowing us to remove
a chunk of custom cleaning code.

Also fixes #469 (changing the sign on the wrong column), which occurred based
on the addition of a new column.