Include `future` module in installation and packaging
Having added `future` for Python 2/3 compatibility we need to automatically
install it when installing via pip and when packaging the retriever.

It had been added to `extra_includes` but in a way that only got included on
Linux. This moves it to `includes` where it will always get included in the
binary packages and adds it to `install_requires` where it will get picked up
when installing via `pip`.