Update location and details of Portal supplemental tables
The supplemental tables (species and plots) for the Portal database aren't
available as part of the core dataset, so they had been stored on the old
Ecological Data Wiki server. We failed to move them in the update of the
data wiki so those links are now failing. I've updated the script to use
the tables in the Portal Project Teaching Database on Figshare, which will
be a more stable location until we update to the newest Portal release.