Get try_install_all running and exclude more datasets
This updates try_install_all to get it running again. There was a very
strange error trigger by the `if value` check for the download only engine.
For some reason the value for the download only engine doesn't evaluate to
True or False when using `if value`.

Also excludes some of the new large datasets.