add get_input function
currently mysql engine is failing due to lack of connection.

since the retriever gets the opts using command line, get_input()
will ask the user for the opts and get_cursor() will use these values
to connect to the various engines.

In case an engine fails to connect we record that as an error as
errors.append((key, "No connection detected"))
if an engine error already exist, we do not add it in the error list

as we run the