improve integration tests and SQlite engine bulk insert.
We have improved  integration tests and have added
integration for all engines.
SQlite engine improvement:
During bulk insert, SQlite included endofline characters
in the row. (1,2,"3\n") instead of (1,2,3).
This PR removes the endofline characters.
This issue was not visible when creating the table.
A select on the created table would reveal the issue.

fixes #310