Add row value length checks before adding to table (with integraton tests)
Some datasets have an extra newline character in their rows, which causes
retriever to read them as separate instances on data, This causes mismatch
int the numbers if columns and data values, producing IndexError exceptions
in a few cases.

This commit adds a loop to keep adding lines to a row until the number of values
equals the number of columns, before adding the row to the table.

Also added integration test 'test_extra_newline'.