Update link to Gentry sites table
This table was digitized from a pdf and posted it to the Ecological Data Wiki.
When the wiki was upgraded earlier this year the table didn't get moved.

I have now posted this table to figshare so that it won't disappear again and
updated the link. The fact that figshare links don't include filenames has been
incorporated into the change.