Use extract_values when checking for mid-row line endings
When handling extra mid-row line endings we split the line to check if there
are enough values. This was added in 9f05ea3, but it ran split_on_delimiter
without checking if the table has a delimiter and so errored on fixed-width
tables. This replaces `split_on_delimiter` with `extract_values` to properly
split the line in both delimited and fixed-width cases.