Only use `extract_values` for splitting lines
`extract_values` properly handles both fixed width and delimited data
and should therefore be used anytime it is necessary to split up rows
of data. This replaces the few instances of `split_on_delimiter` used
outside of `extract_values` with `extract_values` and improves the
documentation for both functions to make clear that `extract_values`
should be used preferentially.