refine conversion of strings to floats in the engine
We did not handle the conversion of strings to floats appropriately.
In some dataset, where null values are not specified,
installing will give some error.
for example, in FrayJorge 'ND' looks like the null value and
since we can not change to float, the system will error,

In the current PR, if we can't change a value to float, we
take that as a null