Edited Developer docs
I changed the developer docs for more clarification on how to do basic setup for the retriever. Since PostgreSQL doesn't run into the same problem, it seemed to have been overlooked, but for MySQL a new user named travis and a new database named testdb must be created for everything to function correctly.

fixes #2