Make sure temporary `raw_data` test directory is removed
Starting with the merge of #626 we create a copy of the `test/raw_data` directory
in the root directory for testing. It wasn't getting properly cleaned up by all,
of the test modules. This fixes that and replaces a couple `os.system("rm -r...")
calls with a better `shutil.rmtree`.