Rearranged the Documents so for uniformity and completeness
I moved the services used section to running test locally. Also, I
deleted the installation instructions for mysql, and added commands in
mysql to remind the reader to create users. Mysql commands also now take
the form " mysql -e """ to keep parallel with the postgresql commands
and so that they can be executed straight from the command line.