Remove urlretrieve import from python 2
Why?
`urlretrieve` function in Python 2 is buggy and returns [Error 200] while downloading archives, maybe because of a socket-closing issue.

Fix:
This issue is not present when using the `future` package, which backports urllib from Python 3.
Using future.standard_library.install_aliases, the Python 3 API with the backports is made available in Python 2.