Use CSV module for parsing lines.
This module reads lines from the source, and we remove extra
line endings in any cell.

This enables processing of the lines correctly for we are working with
values of a list. Each list item corresponds to a given header value.

Processing of the `fixed width files` is not handled by the csv module but
but the function `extract_fixed_width`. Like the CSV module, we return
a list of values that corresponds to the list of columns.

Add test for extra new lines edge case.

Extra new lines are created when there is a line end in a cell.
To view this scenario you can have a look at the EltonTraits
(http://esapubs.org/archive/ecol/E095/178/MamFuncDatSources.txt).
`
Ref_34     "Emmons L.H. & F. Feer. 1997. Neotropical Rain Forest Mammals, a Field Guide. 2da ed. The
University of Chicago Press. 281 pp"
`
The above example could be confused for two lines and this becomes
complicated when more columns are involved.
When opened in txt form, the files rendering fails, however when opened
in CSV applications, The right rendering is obtained.