Changes to new json creation function: added version and url keys, fixed null values
When we create a new json scripts, the version number is automatically added and
set to 1.0 now, and all table urls are compiled and placed in the urls
dictionary at the end of the scripts, just like in other normal scripts.
Null values previously when inputted in the new_json function would
cause an error during compilation. This was remedied in my changes to
compile.py