Fix a data entry/formatting error in Gentry
On line 130 of the CEDRAL data there is an extra ' in the voucher1 column.
This is related to a parsing error in the raw Excel file that makes all
following columns off by one, which results in the Counts column containing
non-integer values (because the first size ends up in the counts column)
and the stems table will be missing a single, 2.5, value.

This fix finds that row and then manually updates the relevant values to the
correct ones.

Fixes #415.