In auto_get_datatypes add 100 when comparing to the max_lengths
Since we decided to add 100 to the detected length of char data types,
comparisions against length should also add 100 to the length of the string(var)

Bug was noticed in baad methods files where field a.shba has a first non empty value of
"?". Raising the number of max_lengths to 101.  However, since this is cleaned up as "None", other values
don't get to set the datatypes since all are of length < 101.