Make script search paths work in correct order
The design is to look for scripts first in ., then in ./scripts, and finally
in ~/.retriever/scripts. The first script found should be loaded and any additional
copies of the same script ignored. The previously implementation looked good, but had
issues with aliasing, so even though copies of scripts from later directories weren't
explicitly added, they were replacing the correct scripts in the list. This separates
the check for whether the script has already been found to avoid the aliasing issue.

This design allows for easy development and customization. When developing modified
scripts in the repository are used by default. When customizing scripts added to
the working directory are used preferentially. When no other scripts appear the core
scripts are used.