Accomodate new retriever naming conventions in fetch
The soon to be released version 2.0 of the core retriever has moved to using dataset
names that include words separated by `-` in line with the datapackage spec. These
names are then converted to `_` for compatibility with some database managment systems.
This resulted in `fetch()` not finding the temporary files it had downloaded because
the dataset name and relevant piece of the file name didn't match precisely. This
converts the dataset name to `_` when matching.

Fixes #75.