Don't change the class or return the update log
Get updates was storing the output of running `retriever update`, setting
it's class to `update_log` and returning this output. The current output
from `retriever update` isn't genearlly useful, it's mostly just a progress
bar, and for some reason changing it's class was creating an error (see #82)
so this just change just prints the third value from `update_log`, which
indicates that the update is complete.

Fixes #82.