Make RStudio on Windows use standard HOME
RStudio on Windows gives HOME paths of the Documents folder in the users
home directory. This is different from all other operating systems and
ways of running R on Windows. This difference cascades through the CLIs
wrapped using R. This checks to see if hte Documents folder is part of
the home directory, removes it, and sets HOME for the R session.

This is based on @henrykironde's debugging of this issue in #90.

Fixes #81.