Adding choices through argcomplete
Three lists of choices have been prepared for argcomplete to work on -
script_list (list of all datasets available), json_list (list of all
datasets available in json format) and keywords_list (list of all the
tags available in the datasets). The choices have been added in the form
of function argument to add_argument function or as completer attribute
(when choices outside the given list is permitted).
  - script_list: This choice has been added for citation, download and
  install command
  - json_list: This choice has been added for edit_json and delete_json
  command
  - keywords_list: This choice has been added specifically for 'ls -l'
  command for listing all datasets based on keywords provided