Addition and changes to the precommit hook
Changed version numbers to use the python module LooseVersion. This
module is useful for comparing version numbers in the form x.y.z and
also realizing that x.y.9 is less than x.y.10

Setup.py now performs a check to see if a pre-commit hook exists first,
and only overwrites if it does not.

version_files compares against last commit version files. This may need
to be discussed further.