Modifying setup.py and requirements to install and setup argcomplete
As refered in the argcomplete documentations, it needs to be enabled for
global completion after installation. Furthermore, retriever needs to be
registered for global completion with argcomplete. These two necessities
have been resolved in this commit. As refered in #815, argcomplete has
been added as an installation requirement. Since argcomplete onl
provides completion for bash, it is only enabled when the current
platform is detected as not to be windows.