Re-add high speed bulk importing for SQLite
This was originally added in bc182e4 and then removed in defdea4 because
we thought all of that functionality had been moved into the main add_to_table
code. However the speed ups dropped off substantially with that move and it turns
out there are some differences in how inserts are handled by this code since it
can assume data that doesn't need to be cleaned (e.g., chunked line reading).

This restores the code to it's previous form so that the bulk inserts work quickly
again. This is on the order of a 20x speed up for large clean datasets.