Check engine arguments properly for tests
When `test_modified_scripts.py` was run as a test and not directly
it wasn't working as expected when arguments are passed to the test
suite. It checked to see if any arguments have been passed and
interpreted them as a subset of engines to test. Since none of the
test suite arguments are engine names the tests tests in this file
aren't run.

This change moves that code in the `if __name__ == "__main__" block
to make sure it is only handled when appropriate.