1.Removes compiling from json to python scripts. 2.Add new tables for spatial data. For the retriever to handle Geospatial data
3.Define table for each script before pre-processing.

4.Added the functionaly to handle
compressed/archied datasets by using key word archieved
and the files that are in the zip for each table.

5. Handling data packages registered in the `datapackage.yml` file
I have remove the scripts from this file that seem
not to have followed the standards.
To make this possible, we are matching the major `frictionless`
types, integer,string etc,  with the types in the retriever and falling
back to char for some other types
For example, datetime YYYY-MM-DDThh:mm:ssZ.

Ref: This is a combination of the work from PR #814 #980 #1005