Removing compilation of python scripts.
1.Removes compiling from json to python scripts.
2.Add new tables for spatial data.
For the retriever to handle Geospatial data

3.Define preprocessing on table rather than script.
Instead of defining the entire functionaly
of preprocessing based on the script,
we define the preprocessing based on tables.

We have given power to tables to demostrate
their functionality and thus making the
script flexible to handle various processing on tables

4.Added the functionaly to handle
compressed/archied datasets by using key word archieved
and the files that are in the zip for each table.