Create gdp.json
The Following pr is with reference with the following issue https://github.com/weecology/retriever/issues/967 . 
There was some problem with the data package as the latest json schema did not support the format provided 
in the datapackage , as a result i have to include all the years from 1960 to 2017 as columns to my table. 
Please verify and get back.