Remove directory name from gentry-forecast-transects site code
At some point a behavior changed that caused site names extracted from the file
names in gentry-forest-transects to including the directory in which they were
located. This leads to site_code values like africa/NDAKANNI. This commit
properly normalizes the path to get only the file name.

Also removes additional printing done in this script that is no longer desired.